requirements.txt
attrs==20.2.0
blinker==1.4
click==7.1.2
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
flake8==3.8.4
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
more-itertools==8.5.0
packaging==20.4
pg8000==1.16.5
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.2
python-dotenv==0.14.0
scramp==1.2.0
six==1.15.0
SQLAlchemy==1.3.19
testing.common.database==2.0.3
testing.postgresql==1.3.0
toml==0.10.1
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
